+++ data/hax/special_spawnables.lua
@@ -1,5 +1,15 @@
 -- special spawnable items that don't get found by `gen_spawnlist.py`
 special_spawnables = {
-  {path='data/entities/animals/boss_alchemist/key.xml', name=tr("spawn.crystal_key"), xml="key.xml"},
-  {path='data/entities/animals/boss_centipede/sampo.xml', name=tr("spawn.sampo"), xml="sampo.xml"},
+  {
+    path='data/entities/animals/boss_alchemist/key.xml',
+    name=tr("spawn.crystal_key"),
+    english_name=tr_english("spawn.crystal_key"),
+    xml="key.xml"
+  },
+  {
+    path='data/entities/animals/boss_centipede/sampo.xml',
+    name=tr("spawn.sampo"),
+    english_name=tr_english("spawn.sampo"),
+    xml="sampo.xml"
+  },
 }